feat(STX-575): move computeHasSufficientGasForQuoteForMetrics to useHasSufficientGasForQuoteForMetrics.ts
